LayoutTests/http/tests/from-origin/document-nested-from-origin-same-accepted-expected.txt

Tests that a nested same-origin document load succeeds if the server blocks cross-origin loads with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
The outer iframe, loading a nested 127.0.0.1 iframe.



--------
Frame: '<!--framePath //<!--frame0-->/<!--frame0-->-->'
--------
The iframe
